Fixing tests and getting ES search to work better with filters.
