+++ backend/node_app/utils/searchUtility.js
@@ -58,7 +58,6 @@ class SearchUtility {
 		this.getSearchCount = this.getSearchCount.bind(this);
 		this.getRecDocs = this.getRecDocs.bind(this);
 		this.getJBookPGQueryAndSearchTerms = this.getJBookPGQueryAndSearchTerms.bind(this);
-		this.getElasticSearchJBookDataFromId = this.getElasticSearchJBookDataFromId.bind(this);
 	}
 
 	createCacheKeyFromOptions({ searchText, cloneName = 'gamechangerDefault', index, cloneSpecificObject = {} }){
@@ -2958,204 +2957,7 @@ class SearchUtility {
 		return query;
 	}
 
-	getElasticSearchQueryForJBook({searchText, parsedQuery, offset, limit, jbookSearchSettings, operator = 'and'}, userId, serviceAgencyMappings) {
 
-		const isVerbatimSearch = this.isVerbatim(searchText);
-		const plainQuery = (isVerbatimSearch  ? parsedQuery.replace(/["']/g, '') : parsedQuery);
-
-		let query = {
-			track_total_hits: true,
-			from: offset,
-			size: limit,
-			aggregations: {
-				service_agency_aggs: {
-					terms: {
-						field: 'serviceAgency_s',
-						size: 10000
-					}
-				}
-			},
-			query: {
-				bool: {
-					must: [],
-					should: [
-						{
-							multi_match: {
-								query: `${parsedQuery}`,
-								fields: esTopLevelFields,
-								type: 'best_fields',
-								operator: `${operator}`
-							}
-						}
-					]
-				}
-			},
-			highlight: {
-				fields: {}
-			}
-		};
-
-		esTopLevelFields.forEach(field => {
-			query.highlight.fields[field] = {};
-		});
-
-		esInnerHitFields.forEach(innerField => {
-			const nested = {
-				nested: {
-					path: innerField.path,
-					inner_hits: {
-						_source: false,
-						highlight: {
-							fields: {}
-						}
-					},
-					query: {
-						bool: {
-							should: [
-								{
-									multi_match: {
-										query: `${parsedQuery}`,
-										fields: innerField.fields,
-										type: 'best_fields',
-										operator: `${operator}`
-									}
-								}
-							]
-						}
-					}
-				}
-			};
-
-			innerField.fields.forEach(field => {
-				nested.nested.inner_hits.highlight.fields[field] = {};
-			});
-			query.query.bool.should.push(nested);
-		});
-
-		const wildcardList = {
-			'type_s': 1,
-			'key_s': 1,
-			'projectTitle_t': 1,
-			'projectNum_s': 6,
-			'programElementTitle_t': 1,
-			'serviceAgency_s': 2,
-			'appropriationTitle_t': 1,
-			'appropriationNumber_s': 6,
-			'budgetActivityTitle_t': 6,
-			'budgetActivityTitle_s': 6,
-			'programElement_s': 6,
-			'accountTitle_s': 1,
-			'budgetLineItemTitle_s': 1,
-			'budgetLineItem_s': 6
-		};
-
-		Object.keys(wildcardList).forEach((wildCardKey) => {
-			query.query.bool.should.push({
-				wildcard: {
-					[wildCardKey]: {
-						value: `*${plainQuery}*`,
-						boost: wildcardList[wildCardKey]
-					}
-				}
-			});
-		});
-
-		// FILTERS
-
-		console.log(jbookSearchSettings);
-
-		if (jbookSearchSettings && jbookSearchSettings.budgetYear) {
-			query.query.bool.must.push({
-				terms: {
-					budgetYear_s: jbookSearchSettings.budgetYear
-				}
-			});
-		}
-
-		if (jbookSearchSettings && jbookSearchSettings.serviceAgency) {
-			const convertedAgencies = [];
-
-			jbookSearchSettings.serviceAgency.forEach(agency => {
-				Object.keys(serviceAgencyMappings).forEach(agencyKey => {
-					if (serviceAgencyMappings[agencyKey] === agency) {
-						convertedAgencies.push(agencyKey);
-					}
-				});
-			});
-			query.query.bool.must.push({
-				terms: {
-					serviceAgency_s: convertedAgencies
-				}
-			});
-		}
-
-		if (jbookSearchSettings && jbookSearchSettings.budgetType) {
-			const budgetTypesTemp = [];
-
-			jbookSearchSettings.budgetType.forEach(budgetType => {
-				switch (budgetType) {
-					case 'RDT&E':
-						budgetTypesTemp.push('rdte');
-						break;
-					case 'O&M':
-						budgetTypesTemp.push('om');
-						break;
-					case 'Procurement':
-						budgetTypesTemp.push('procurement');
-						break;
-					default:
-						break;
-				}
-			});
-
-			query.query.bool.must.push({
-				terms: {
-					type_s: budgetTypesTemp
-				}
-			});
-		}
-
-		// SORT
-		switch (jbookSearchSettings.sort[0].id) {
-			case 'budgetYear':
-				query.sort = [{budgetYear_s: {order: jbookSearchSettings.sort[0].desc ? 'desc' : 'asc'}}];
-				break;
-			case 'programElement':
-				query.sort = [{	programElement_s: {order: jbookSearchSettings.sort[0].desc ? 'desc' : 'asc'}}];
-				break;
-			case 'projectNum':
-				query.sort = [{	projectNum_s: {order: jbookSearchSettings.sort[0].desc ? 'desc' : 'asc'}}];
-				break;
-			case 'projectTitle':
-				query.sort = [{	projectTitle_s: {order: jbookSearchSettings.sort[0].desc ? 'desc' : 'asc'}}];
-				break;
-			case 'serviceAgency':
-				query.sort = [{	serviceAgency_s: {order: jbookSearchSettings.sort[0].desc ? 'desc' : 'asc'}}];
-				break;
-			default:
-				break;
-		}
-
-		return query;
-	}
-
-	getElasticSearchJBookDataFromId({docIds}, userId) {
-		try {
-			return {
-				track_total_hits: true,
-				size: 100,
-				query: {
-					bool: {
-						must: {
-							terms: {key_review_s: docIds}
-						}
-					}
-				}
-			};
-		} catch (err) {
-			this.logger.error(err, '1F07MYM', userId);
-		}
-	}
 
 }
